requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.0
openai>=1.0.0
ollama>=0.2.0
matplotlib>=3.8.0
graphviz>=0.20.0
pandas>=2.2.0
